Add basic logging option, fix settings docstring
Add option to run connection script "(with logging)", which
outputs the user's config, environment, and verbose output
of the openconnect binary to $HOME/Library/Logs/xbar-openconnect.log.
Note that this location is available via the mac Console app.

The docstring had the binary locations for the two
architectures (Intel and Apple) reversed.